Refactor to classes
